+++ CHANGELOG.md
@@ -65,6 +65,11 @@
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#1371](https://github.com/realm/SwiftLint/issues/1371)
 
+* Add `collection_alignment` opt-in rule to validate that all elements in a
+  collection literal are aligned vertically.  
+  [Dalton Claybrook](https://github.com/daltonclaybrook)
+  [#2326](https://github.com/realm/SwiftLint/issues/2326)
+
 #### Bug Fixes
 
 * Fix `comma` rule false positives on object literals (for example, images).